feat: make it possible to delete parts, chapters and lessons
