+++ extensions/vscode/src/commands/tutorialkit.load-tutorial.ts
@@ -1,21 +1,22 @@
 import * as vscode from 'vscode';
 import { extContext } from '../extension';
-import { LessonsTreeDataProvider, setLessonsTreeDataProvider } from '../views/lessonsTree';
+import { LessonsTreeDataProvider } from '../views/lessonsTree';
+import { setLessonsTreeDataProvider, setLessonsTreeView } from '../global-state';
 
 export async function loadTutorial(uri: vscode.Uri) {
   const treeDataProvider = new LessonsTreeDataProvider(uri, extContext);
 
   await treeDataProvider.init();
 
-  setLessonsTreeDataProvider(treeDataProvider);
-
-  extContext.subscriptions.push(
-    vscode.window.createTreeView('tutorialkit-lessons-tree', {
-      treeDataProvider,
-      canSelectMany: true,
-    }),
+  const treeView = vscode.window.createTreeView('tutorialkit-lessons-tree', {
     treeDataProvider,
-  );
+    canSelectMany: true,
+  });
+
+  setLessonsTreeDataProvider(treeDataProvider);
+  setLessonsTreeView(treeView);
+
+  extContext.subscriptions.push(treeView, treeDataProvider);
 
   vscode.commands.executeCommand('setContext', 'tutorialkit:tree', true);
 }
